Reduce precision in GNN

+++ models/non_autoreg.py
@@ -62,10 +62,11 @@ class NonAutoRegressive(nn.Module):
         if self.encode == 'sequence' and use_reads:
             h = self.seq_encoder(reads)
         elif self.encode == 'node':
-            h = torch.ones((graph.num_nodes(), 1)).to(self.hyperparams['device'])
+            h = torch.ones((graph.num_nodes(), 1), dtype=torch.float16).to(self.hyperparams['device'])
             h = self.node_encoder(h)
         else:
-            h = torch.ones((graph.num_nodes(), self.hyperparams['dim_latent'])).to(self.hyperparams['device'])
+            h = torch.ones((graph.num_nodes(), self.hyperparams['dim_latent']), dtype=torch.float16).to(self.hyperparams['device'])
+        h = h.type(torch.float16)
 
         # norm = self.hyperparams['norm']
         if norm is not None:
@@ -74,6 +75,7 @@ class NonAutoRegressive(nn.Module):
             e_tmp = graph.edata['overlap_length'].float() 
             e_tmp = (e_tmp - torch.mean(e_tmp)) / torch.std(e_tmp)
         e = self.edge_encoder(graph.edata['overlap_similarity'], e_tmp)
+        e = e.type(torch.float16)
         e_f = e.clone()
         e_b = e.clone()
         for conv in self.layers:
